modify throttle wait params type to double

+++ src/main/kotlin/io/github/mainyf/minecraftlib/IMinecraftLib.kt
@@ -12,6 +12,6 @@ interface IMinecraftLib {
     /**
      * get throttle tool class
      */
-    fun getThrottle(wait: Long): Throttle
+    fun getThrottle(wait: Double): Throttle
 
 }
